Update README.md
Following the commands in the contribution guide does not build the local
development environment correctly. First, email_validator is also required
as python module, but not listed on the requirements.txt. Second, it is better
to use `pip install -r requirements.txt` instead of installing the required
modules one by one manually.