Prepare Docker image for released pervane versions
Pervane can be used with Docker. Users can pull the pervane/pervane
docker image and start containers with it.

How to build and how to use guides are available under docker/release
and they will be integrated to the documentation once we decide on how
to structure the documentation on Wiki.